+++ src/behaviour/behaviour.js
@@ -1,0 +1,4 @@
+export default (applyBehaviour) => (h, {skin}) => (props, children) => {
+  applyBehaviour(props, children, skin);
+  return children;
+}
