--- src/colors.css
@@ -1,1 +1,0 @@
-@value primary: #3498db;
