+++ src/index.js
@@ -1,5 +1,5 @@
-import createSlide from './Slide';
-import createTitle from './Slide/Title';
+import createSlide from './organism/slide';
+import createTitle from './atom/title';
 
 export {
     createSlide,
