+++ src/style/colors.css
@@ -1,0 +1,2 @@
+@value primary: #12F;
+@value secondary: #1b4;
